Update README.md
For this assignment I was asked to use D3 data visualisation to represent the data taken from the vehicles.csv file supplied by the central statistics office (CSO.IE). The file contains details of 
vehicles registered in Ireland over the past five years.

I was tasked with creating a web page that has the data visualisation capabilities to represent this data in the following 
visual ways: line or scatter plot, bar, pie and doughnut charts. The charts are required to represent the following:
• graph displaying all new registered vehicles past five years.
• graph for displaying all registered vehicles in a selected year.
• graph highlighting the differences in registration of private and second-hand vehicles over 
the past five years. (see CSO example diagram below).
• graph of tractor registrations per month.
• student data visualization of the vehicle data, the information chosen is left to your own discretion.